refactoring: TabIcons의 복수개의 컴포넌트를 파일로 분리함
PollTabIcon, QnATabIcon으로 파일 분리
