--- frontend/guest-app/src/components/TabGroup/TabIcons.js
@@ -1,28 +1,0 @@
-import React from "react";
-import Badge from "@material-ui/core/Badge/Badge.js";
-import QuestionAnswerIcon from "@material-ui/icons/QuestionAnswer";
-import PollIcon from "@material-ui/icons/Poll";
-
-const style = {marginRight: "8px"};
-
-export function QnATabIcon({showBadge}) {
-	const props = showBadge ? {color: "error", variant: "dot"} : {};
-
-	return (
-		<Badge {...props}>
-			<QuestionAnswerIcon style={style} />
-			Q&A
-		</Badge>
-	);
-}
-
-export function PollTabIcon({showBadge}) {
-	const props = showBadge ? {color: "error", variant: "dot"} : {};
-
-	return (
-		<Badge {...props}>
-			<PollIcon style={style} />
-			투표
-		</Badge>
-	);
-}
